pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-env-proxy"
version = "0.1.0"
authors = [
    { name = "Kamorion Labs", email = "contact@kamorion.cloud" },
]
description = "A dynamic MCP proxy that allows switching environment variables on-the-fly"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[project.scripts]
mcp-env-proxy = "mcp_env_proxy.__main__:main"

[project.urls]
Homepage = "https://github.com/KamorionLabs/mcp-env-proxy"
Repository = "https://github.com/KamorionLabs/mcp-env-proxy.git"
Issues = "https://github.com/KamorionLabs/mcp-env-proxy/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_env_proxy"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
